fix: flatten protocol_room folder structure
