+++ lichen-protocol-mvp/rooms/protocol_room/integrity.py
@@ -4,7 +4,7 @@ Implements Integrity Gate theme from Protocol Room Protocol
 """
 
 from typing import List
-from .room_types import IntegrityResult
+from .types import IntegrityResult
 
 
 def check_stones_alignment(protocol_text: str) -> bool:
@@ -19,14 +19,13 @@ def check_stones_alignment(protocol_text: str) -> bool:
     integrity_indicators = [
         "integrity", "honesty", "authenticity", "truth",
         "clarity", "simplicity", "directness",
-        "care", "respect", "safety", "trust",
-        "help", "support", "guide", "assist", "enable"
+        "care", "respect", "safety", "trust"
     ]
     
     # Check for potential misalignment indicators
     misalignment_indicators = [
         "manipulation", "deception", "complexity", "confusion",
-        "pressure", "urgency", "demand", "force", "coerce"
+        "pressure", "urgency", "demand"
     ]
     
     # Count positive indicators
@@ -35,12 +34,8 @@ def check_stones_alignment(protocol_text: str) -> bool:
     # Count negative indicators
     negative_count = sum(1 for indicator in misalignment_indicators if indicator in text_lower)
     
-    # More lenient scoring: protocols are aligned by default unless they have clear misalignment
-    # Give a baseline score for well-structured protocols
-    baseline_score = 2 if any(word in text_lower for word in ['purpose', 'step', 'help', 'guide']) else 0
-    
-    total_score = positive_count + baseline_score
-    return total_score >= negative_count
+    # Simple scoring: more positive than negative = aligned
+    return positive_count >= negative_count
 
 
 def check_coherence(protocol_text: str) -> bool:
@@ -51,24 +46,19 @@ def check_coherence(protocol_text: str) -> bool:
     # Simple deterministic checks
     text_lines = protocol_text.split('\n')
     
-    # Check for reasonable length - be more lenient
-    reasonable_length = 20 <= len(protocol_text) <= 10000
+    # Check for basic structure
+    has_title = any(line.startswith('#') for line in text_lines)
+    has_steps = any('step' in line.lower() or '1.' in line or '2.' in line for line in text_lines)
+    has_completion = any('completion' in line.lower() for line in text_lines)
     
-    # Check for clear language - be more lenient
+    # Check for reasonable length
+    reasonable_length = 50 <= len(protocol_text) <= 5000
+    
+    # Check for clear language
     clear_language = not any(word in protocol_text.lower() for word in ['jargon', 'confusing', 'unclear'])
     
-    # For scenario text (single line), just check length and clarity
-    if len(text_lines) <= 2:
-        # But still check for some basic structure indicators
-        has_basic_structure = any(word in protocol_text.lower() for word in ['purpose', 'step', 'help', 'guide', 'when', 'take'])
-        return reasonable_length and clear_language and has_basic_structure
-    
-    # For full protocols, check for structure
-    has_title = any(line.startswith('#') for line in text_lines)
-    has_steps = any('step' in line.lower() or '1.' in line or '2.' in line or '**' in line for line in text_lines)
-    
-    # Must have basic structure elements for full protocols
-    return has_title and has_steps and reasonable_length and clear_language
+    # Must have basic structure elements
+    return has_title and has_steps and has_completion and reasonable_length and clear_language
 
 
 def run_integrity_gate(protocol_text: str) -> IntegrityResult:
