--- lichen-protocol-mvp/rooms/protocol_room/protocol_room/__init__.py
@@ -1,25 +1,0 @@
-"""
-Protocol Room Module
-Implements the Protocol Room Protocol and Contract for Lichen Protocol Room Architecture (PRA)
-"""
-
-from .protocol_room import ProtocolRoom, run_protocol_room
-from .types import (
-    ProtocolRoomInput,
-    ProtocolRoomOutput,
-    ProtocolDepth,
-    ProtocolText,
-    ScenarioMapping,
-    IntegrityResult
-)
-
-__all__ = [
-    'ProtocolRoom',
-    'run_protocol_room',
-    'ProtocolRoomInput',
-    'ProtocolRoomOutput',
-    'ProtocolDepth',
-    'ProtocolText',
-    'ScenarioMapping',
-    'IntegrityResult'
-]
